# openapi.yaml
openapi: 3.0.0
info:
  title: SportsDeck API
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
...
  /api/posts/{id}/replies:
    post:
      summary: Create a reply under a post
      tags:
        - Posts
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          example: clx1abc123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
                  example: This is my reply to the post.
      responses:
        '201':
          description: Reply created successfully
        '400':
          description: Content is required
        '401':
          description: Unauthorized
        '403':
          description: Account banned or thread locked/closed
        '404':
          description: Post or thread not found
        '500':
          description: Internal server error
    get:
      summary: Get all visible replies under a post
      tags:
        - Posts
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          example: clx1abc123
      responses:
        '200':
          description: List of replies
        '404':
          description: Post not found
        '500':
          description: Internal server error
  /api/posts/{id}:
    patch:
      summary: Edit a post's content (owner only)
      tags:
        - Posts
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          example: clx1abc123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
                  example: Updated post content here.
      responses:
        '200':
          description: Post updated successfully
        '400':
          description: Content is required
        '401':
          description: Unauthorized
        '403':
          description: Forbidden or account banned or content hidden
        '404':
          description: Post not found
        '500':
          description: Internal server error
    delete:
      summary: Soft-delete a post (owner or admin)
      tags:
        - Posts
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          example: clx1abc123
      responses:
        '200':
          description: Post deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Post not found
...
  /api/replies/{id}:
    patch:
      summary: Edit a reply's content (owner only)
      tags:
        - Replies
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          example: clx2reply456
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
                  example: Updated reply content.
      responses:
        '200':
          description: Reply updated successfully
        '400':
          description: Content is required
        '401':
          description: Unauthorized
        '403':
          description: Forbidden or account banned or content hidden
        '404':
          description: Reply not found
        '500':
          description: Internal server error
    delete:
      summary: Soft-delete a reply (owner or admin)
      tags:
        - Replies
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          example: clx2reply456
      responses:
        '200':
          description: Reply deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Reply not found
...